requirements.txt
fastapi==0.115.12
uvicorn[standard]==0.34.2
sqlalchemy==2.0.40
pydantic==2.11.3
pydantic-settings==2.8.1
python-multipart==0.0.20
python-jose[cryptography]==3.4.0
passlib[bcrypt]==1.7.4
jinja2==3.1.6
httpx==0.28.1
openai>=1.0.0
google-genai>=1.0.0
email-validator>=2.2.0
pytest==8.3.5
